requirements.txt
asgiref==3.2.7
astroid==2.3.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
google-auth==1.13.1
gspread==3.4.2
httplib2==0.17.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.2
oauth2client==4.1.3
pandas==1.0.3
Pillow==7.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2